+++ pyticketswitch/availability.py
@@ -118,7 +118,8 @@ class AvailabilityDetails(JSONMixin, object):
                  last_date=None, calendar_masks=None,
                  weekday_mask=None, valid_quantities=None,
                  cached_number_available=None, discount_code=None,
-                 discount_desc=None, discount_semantic_type=None):
+                 discount_desc=None, discount_semantic_type=None,
+                 suffixed_price_band_code=None):
 
         self.ticket_type = ticket_type
         self.ticket_type_description = ticket_type_description
@@ -140,6 +141,7 @@ class AvailabilityDetails(JSONMixin, object):
         self.discount_code = discount_code
         self.discount_desc = discount_desc
         self.discount_semantic_type = discount_semantic_type
+        self.suffixed_price_band_code = suffixed_price_band_code
 
     @classmethod
     def from_api_data(cls, data):
@@ -185,6 +187,8 @@ class AvailabilityDetails(JSONMixin, object):
                     kwargs['discount_desc'] = raw_details.get('discount_desc')
                     kwargs['discount_semantic_type'] = raw_details.get(
                         'discount_semantic_type')
+                    kwargs['suffixed_price_band_code'] = raw_details.get(
+                        'suffixed_price_band_code')
 
                     available_dates = raw_details.get('available_dates', {})
                     if 'first_yyyymmdd' in available_dates: